server/requirements.txt
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0 
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2==2.9.3
pycparser==2.21
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.1
PyJWT==1.7.1

